Update Parser.java
Didn't want to touch alot the exposed interface but this getter setter of the file really disturbed in executing my plan to make it threadsafe. 
File is always a leak, but at least we should not encourage it by contract.
changes are also commented inline. Here the succus:
(1) imutable file, (2) internal DRY method for content reading, then (3) locking it was more efficient. Could have used a sync(lock){} mutex logic but preferred the more semantically explicit read and write lock logic of ReentrantReadWriteLock.
(4) not clear why its called getContentWithoutUnicode was, because a quick check shows that you check for Unicode Block Basic Latin